Create LICENSE
Does adding one now even do anything. People who stole anything stole nothing but what is still only a half-baked prototype of an idea which itself is founded on my own shaky understanding of how this will all play out. Good luck to anyone who wants to try and make this vision a reality, but please don't be a dick.